+++ bloom_nofos/users/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
+from django.utils.timezone import localtime
 
 from .forms import BloomUserChangeForm, BloomUserCreationForm
 from .models import BloomUser
@@ -13,10 +14,37 @@ class BloomUserAdmin(UserAdmin):
         "email",
         "full_name",
         "group",
-        "is_staff",
+        "is_staff_status",
         "is_active",
-        "is_superuser",
+        "is_superuser_status",
+        "formatted_last_login",
     )
+
+    def formatted_last_login(self, obj):
+        # Format the last_login datetime in a custom format
+        if obj.last_login:
+            # Ensure the datetime is timezone-aware and localized
+            local_last_login = localtime(obj.last_login)
+            return local_last_login.strftime("%d\u00A0%b,\u00A0%H:%M")
+        return "—"
+
+    formatted_last_login.short_description = "Last Login"
+    formatted_last_login.admin_order_field = "last_login"
+
+    ## Change is_superuser label
+    def is_superuser_status(self, obj):
+        return obj.is_superuser
+
+    is_superuser_status.short_description = "Superuser"
+    is_superuser_status.boolean = True
+
+    ## Change is_staff label
+    def is_staff_status(self, obj):
+        return obj.is_staff
+
+    is_staff_status.short_description = "Staff"
+    is_staff_status.boolean = True
+
     list_filter = ()
     fieldsets = (
         (None, {"fields": ("email", "full_name", "group", "password")}),