+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Filter the available THEMES by a user's group
 - Filter the available NOFO DESIGNERS by a user's group
 - Admins can now clone a NOFO
+- Show last login time for user accounts on admin screen
 
 ### Changed